// frontend/src/lib/api/generated/types/CreateRestaurantAreaRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { RestaurantAreaType } from "./RestaurantAreaType.ts";

export interface CreateRestaurantAreaRequest {
  /**
   * @type string, uuid
   */
  branchId: string;
  /**
   * @type string, uuid
   */
  floorId: string;
  /**
   * @type string
   */
  name: string;
  /**
   * @type integer
   */
  type: RestaurantAreaType;
  /**
   * @type integer, int32
   */
  sortOrder: number;
}