// frontend/src/lib/api/generated/types/CreateRestaurantFloorPlanRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export interface CreateRestaurantFloorPlanRequest {
  /**
   * @type string, uuid
   */
  branchId: string;
  /**
   * @type string, uuid
   */
  floorId: string;
  /**
   * @type string
   */
  name: string;
  /**
   * @type number, double
   */
  canvasWidth: number;
  /**
   * @type number, double
   */
  canvasHeight: number;
  /**
   * @type null,number, double
   */
  gridSize: number | null;
  /**
   * @type boolean
   */
  isActive: boolean;
}