// frontend/src/lib/api/generated/types/CreateRestaurantReservationEndpointRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { RestaurantReservationSource } from "./RestaurantReservationSource.ts";

export interface CreateRestaurantReservationEndpointRequest {
  /**
   * @type string | undefined, uuid
   */
  branchId?: string;
  /**
   * @type array | undefined
   */
  tableIds?: string[];
  /**
   * @type integer | undefined, int32
   */
  partySize?: number;
  /**
   * @type string | undefined, date-time
   */
  startAt?: string;
  /**
   * @type null,integer, int32
   */
  durationMinutes?: number | null;
  /**
   * @type string | undefined
   */
  customerFullName?: string;
  /**
   * @type null,string
   */
  customerPhone?: string | null;
  /**
   * @type null,string
   */
  customerEmail?: string | null;
  /**
   * @type integer | undefined
   */
  source?: RestaurantReservationSource;
  /**
   * @type null,string
   */
  specialRequests?: string | null;
}