// frontend/src/lib/api/generated/types/CreateRestaurantTableBlockEndpointRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export interface CreateRestaurantTableBlockEndpointRequest {
  /**
   * @type string | undefined, uuid
   */
  branchId?: string;
  /**
   * @type null,string, uuid
   */
  floorId?: string | null;
  /**
   * @type null,string, uuid
   */
  areaId?: string | null;
  /**
   * @type null,string, uuid
   */
  tableId?: string | null;
  /**
   * @type string | undefined, date-time
   */
  startAt?: string;
  /**
   * @type string | undefined, date-time
   */
  endAt?: string;
  /**
   * @type null,string
   */
  reason?: string | null;
}