// frontend/src/lib/api/generated/types/CreateRestaurantTableRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { RestaurantTableShape } from "./RestaurantTableShape.ts";

export interface CreateRestaurantTableRequest {
  /**
   * @type string, uuid
   */
  branchId: string;
  /**
   * @type string, uuid
   */
  floorId: string;
  /**
   * @type null,string, uuid
   */
  areaId: string | null;
  /**
   * @type string
   */
  label: string;
  /**
   * @type integer, int32
   */
  minCapacity: number;
  /**
   * @type integer, int32
   */
  maxCapacity: number;
  /**
   * @type null,integer, int32
   */
  defaultReservationMinutes: number | null;
  /**
   * @type integer
   */
  shape: RestaurantTableShape;
}